+++ scripts/use-packages/use-packages.ts
@@ -14,6 +14,7 @@ import { LOOP_LIMIT } from '../../src/config';
 import { $execCommand } from '../../src/utils/execCommand/$execCommand';
 import { isFileExisting } from '../../src/utils/files/isFileExisting';
 import { commit } from '../utils/autocommit/commit';
+import { assertsError } from '../../src/errors/assertsError';
 
 if (process.cwd() !== join(__dirname, '../..')) {
     console.error(colors.red(`CWD must be root of the project`));
@@ -24,7 +25,8 @@ const program = new commander.Command();
 program.parse(process.argv);
 
 usePackages()
-    .catch((error: Error) => {
+    .catch((error) => {
+        assertsError(error);
         console.error(colors.bgRed(`${error.name} in ${basename(__filename)}`));
         console.error(colors.red(error.stack || error.message));
         process.exit(1);