+++ src/commands/EXPECT/expectCommandParser.ts
@@ -1,6 +1,7 @@
 import spaceTrim from 'spacetrim';
 import { NotYetImplementedError } from '../../errors/NotYetImplementedError';
 import { ParseError } from '../../errors/ParseError';
+import { assertsError } from '../../errors/assertsError';
 import type { ExpectationUnit } from '../../pipeline/PipelineJson/Expectations';
 import { EXPECTATION_UNITS } from '../../pipeline/PipelineJson/Expectations';
 import type { string_markdown_text } from '../../types/typeAliases';
@@ -12,6 +13,8 @@ import type { PipelineTaskCommandParser } from '../_common/types/CommandParser';
 import type { ExpectCommand } from './ExpectCommand';
 
 /**
+import { WrappedError } from '../../errors/WrappedError';
+import { assertsError } from '../../errors/assertsError';
  * Parses the expect command
  *
  * @see `documentationUrl` for more details
@@ -112,9 +115,7 @@ export const expectCommandParser: PipelineTaskCommandParser<ExpectCommand> = {
                 amount,
             } satisfies ExpectCommand;
         } catch (error) {
-            if (!(error instanceof Error)) {
-                throw error;
-            }
+            assertsError(error);
 
             throw new ParseError(
                 spaceTrim(