+++ src/formats/json/utils/isValidJsonString.ts
@@ -1,6 +1,8 @@
+import { assertsError } from '../../../errors/assertsError';
+
 /**
  * Function isValidJsonString will tell you if the string is valid JSON or not
- * 
+ *
  * @public exported from `@promptbook/utils`
  */
 export function isValidJsonString(value: string /* <- [👨‍⚖️] */): boolean {
@@ -8,9 +10,7 @@ export function isValidJsonString(value: string /* <- [👨‍⚖️] */): boole
         JSON.parse(value);
         return true;
     } catch (error) {
-        if (!(error instanceof Error)) {
-            throw error;
-        }
+        assertsError(error);
 
         if (error.message.includes('Unexpected token')) {
             return false;