+++ src/remote-server/startRemoteServer.ts
@@ -8,6 +8,7 @@ import swaggerUi from 'swagger-ui-express';
 import { forTime } from 'waitasecond';
 import { promptbookFetch } from '../_packages/core.index';
 import { CLAIM, DEFAULT_IS_VERBOSE } from '../config';
+import { assertsError } from '../errors/assertsError';
 import { AuthenticationError } from '../errors/AuthenticationError';
 import { PipelineExecutionError } from '../errors/PipelineExecutionError';
 import { serializeError } from '../errors/utils/serializeError';
@@ -327,9 +328,7 @@ export function startRemoteServer<TCustomOptions = undefined>(
             } satisfies LoginResponse<really_any>);
             return;
         } catch (error) {
-            if (!(error instanceof Error)) {
-                throw error;
-            }
+            assertsError(error);
 
             if (error instanceof AuthenticationError) {
                 response.status(401).send({
@@ -427,9 +426,7 @@ export function startRemoteServer<TCustomOptions = undefined>(
                 )
                 .send(source.content);
         } catch (error) {
-            if (!(error instanceof Error)) {
-                throw error;
-            }
+            assertsError(error);
 
             response
                 .status(
@@ -603,9 +600,7 @@ export function startRemoteServer<TCustomOptions = undefined>(
             // <- TODO: [🧠] Should be here transferred data as POST / PUT
             */
         } catch (error) {
-            if (!(error instanceof Error)) {
-                throw error;
-            }
+            assertsError(error);
 
             response.status(400).send({ error: serializeError(error) });
         }
@@ -691,9 +686,7 @@ export function startRemoteServer<TCustomOptions = undefined>(
                     { promptResult } satisfies PromptbookServer_Prompt_Response /* <- Note: [🤛] */,
                 );
             } catch (error) {
-                if (!(error instanceof Error)) {
-                    throw error;
-                }
+                assertsError(error);
 
                 socket.emit('error', serializeError(error) satisfies PromptbookServer_Error /* <- Note: [🤛] */);
             } finally {
@@ -723,9 +716,7 @@ export function startRemoteServer<TCustomOptions = undefined>(
                     { models } satisfies PromptbookServer_ListModels_Response /* <- Note: [🤛] */,
                 );
             } catch (error) {
-                if (!(error instanceof Error)) {
-                    throw error;
-                }
+                assertsError(error);
 
                 socket.emit('error', serializeError(error) satisfies PromptbookServer_Error);
             } finally {
@@ -756,9 +747,7 @@ export function startRemoteServer<TCustomOptions = undefined>(
                         { preparedPipeline } satisfies PromptbookServer_PreparePipeline_Response /* <- Note: [🤛] */,
                     );
                 } catch (error) {
-                    if (!(error instanceof Error)) {
-                        throw error;
-                    }
+                    assertsError(error);
 
                     socket.emit('error', serializeError(error) satisfies PromptbookServer_Error);
                     // <- TODO: [🚋] There is a problem with the remote server handling errors and sending them back to the client